Rename shared/module/module.py to shared/module/utils.py
