+++ rl_algo_impls/shared/actor/categorical.py
@@ -5,7 +5,7 @@ import torch.nn as nn
 from torch.distributions import Categorical
 
 from rl_algo_impls.shared.actor import Actor, PiForward
-from rl_algo_impls.shared.module.module import mlp
+from rl_algo_impls.shared.module.utils import mlp
 
 
 class MaskedCategorical(Categorical):
