+++ rl_algo_impls/shared/encoder/nature_cnn.py
@@ -4,7 +4,7 @@ import gym
 import torch.nn as nn
 
 from rl_algo_impls.shared.encoder.cnn import FlattenedCnnEncoder
-from rl_algo_impls.shared.module.module import layer_init
+from rl_algo_impls.shared.module.utils import layer_init
 
 
 class NatureCnn(FlattenedCnnEncoder):
